pyproject.toml
[project]
name = "wp-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.2.0",
    "duckdb>=1.4.1",
    "matplotlib==3.10.0",
    "numpy>=2.3.3",
    "openai>=2.5.0",
    "pandas==2.2.3",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
    "pyvis>=0.3.2",
    "requests==2.32.3",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
]

[tool.flake8]
exclude = [".venv", "wme_sdk"]
max-line-length = 120

[dependency-groups]
dev = [
    "black>=25.11.0",
    "flake8-pyproject>=1.2.3",
    "ruff>=0.14.5",
    "vulture>=2.14",
]

[tool.ruff]
exclude = [
    ".venv", "wme_sdk"
]
line-length = 120

